+++ example/lib/helpers/fade_route.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class FadeRoute<T> extends MaterialPageRoute<T> {
+  FadeRoute({ WidgetBuilder builder, RouteSettings settings })
+      : super(builder: builder, settings: settings);
+
+  @override
+  Widget buildTransitions(BuildContext context,
+      Animation<double> animation,
+      Animation<double> secondaryAnimation,
+      Widget child) {
+    if (settings.isInitialRoute)
+      return child;
+    return new FadeTransition(opacity: animation, child: child);
+  }
+}
+
+class SimpleFadeRoute<T> extends FadeRoute<T> {
+  SimpleFadeRoute({ Widget child, RouteSettings settings })
+      : super(builder: (_) => child, settings: settings);
+}